requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
langgraph
chromadb
pypdf
tiktoken
python-dotenv
sentence-transformers
langchain-huggingface
